* commit '2fb1fa6c88901fd16f1756a744a5f066f95c9253':
  AArch64: Make AssetAtlasService 64-bit compatible

+++ core/java/android/view/GLRenderer.java
@@ -271,7 +271,7 @@ public class GLRenderer extends HardwareRenderer {
             if (atlas.isCompatible(android.os.Process.myPpid())) {
                 GraphicBuffer buffer = atlas.getBuffer();
                 if (buffer != null) {
-                    int[] map = atlas.getMap();
+                    long[] map = atlas.getMap();
                     if (map != null) {
                         GLES20Canvas.initAtlas(buffer, map);
                     }
